feat : Add scheduler method

+++ Back/sixback/src/main/java/com/a203/sixback/db/repo/PlayerRepo.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PlayerRepo extends JpaRepository<Player, Integer> {
+public interface PlayerRepo extends JpaRepository<Player, Long> {
     List<Player> findAllByTeam_Id(int teamId);
 }
